Split out tax_unit into multiple files
