+++ openfisca_us/variables/demographic/tax_unit/blind_spouse.py
@@ -1,0 +1,9 @@
+from openfisca_us.model_api import *
+
+
+class blind_spouse(Variable):
+    value_type = bool
+    entity = TaxUnit
+    definition_period = YEAR
+    label = "Tax unit spouse is blind"
+    documentation = "1 if spouse is blind; otherwise 0"
