+++ openfisca_us/variables/demographic/tax_unit/mars.py
@@ -22,7 +22,7 @@ class mars(Variable):
         spouse_with_age = person("is_tax_unit_spouse", period)
         has_age = person("age", period) > 0
         has_spouse_with_age = tax_unit.any(spouse_with_age & has_age)
-        has_dependents = add(tax_unit, period, ["is_tax_unit_dependent"]) > 0
+        has_dependents = tax_unit("tax_unit_dependents", period) > 0
         return select(
             [has_spouse_with_age, has_dependents, True],
             [MARSType.JOINT, MARSType.HEAD_OF_HOUSEHOLD, MARSType.SINGLE],
