+++ openfisca_us/variables/demographic/tax_unit/midr.py
@@ -1,0 +1,11 @@
+from openfisca_us.model_api import *
+
+
+class midr(Variable):
+    value_type = bool
+    entity = TaxUnit
+    definition_period = YEAR
+    label = "Separate filer itemizes"
+    documentation = (
+        "True if separately-filing spouse itemizes, otherwise false"
+    )
