+++ openfisca_us/variables/demographic/tax_unit/xtot.py
@@ -1,0 +1,9 @@
+from openfisca_us.model_api import *
+
+
+class xtot(Variable):
+    value_type = int
+    entity = TaxUnit
+    definition_period = YEAR
+    label = "Filing unit exemptions"
+    documentation = "Total number of exemptions for filing unit"
